gc.c: suppress warning
* gc.c (obj_free): suppress a false shorten-64-to-32 warning,
  RUBY_TYPED_FREE_IMMEDIATELY never exceed the limit of int.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@43516 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ gc.c
@@ -1358,7 +1358,7 @@ obj_free(rb_objspace_t *objspace, VALUE obj)
 	    int free_immediately = 0;
 
 	    if (RTYPEDDATA_P(obj)) {
-		free_immediately = RANY(obj)->as.typeddata.type->flags & RUBY_TYPED_FREE_IMMEDIATELY;
+		free_immediately = (int)(RANY(obj)->as.typeddata.type->flags & RUBY_TYPED_FREE_IMMEDIATELY);
 		RDATA(obj)->dfree = RANY(obj)->as.typeddata.type->function.dfree;
 		if (0 && free_immediately == 0) /* to expose non-free-immediate T_DATA */
 		  fprintf(stderr, "not immediate -> %s\n", RANY(obj)->as.typeddata.type->wrap_struct_name);
